requirements.txt
numpy==1.26.4
opencv-python==4.12.0.88
pillow==11.3.0
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
streamlit==1.22.0
streamlit-drawable-canvas==0.8.0
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
albumentations==0.5.2
scikit-image==0.25.2
easydict==1.9
omegaconf==2.1.2
hydra-core==1.1.0
imageio==2.37.0
